+++ src/components/NoteGrid.tsx
@@ -27,6 +27,7 @@ export type NoteGridProps = {
   noteTitleFont: NoteFontPreference;
   noteContentFontSize: NoteFontSize;
   noteTitleFontSize: NoteFontSize;
+  richTextEditorEnabled: boolean;
   noteSize: NoteSizePreference;
   showNoteTitles: boolean;
   showNoteFooters: boolean;
@@ -158,6 +159,7 @@ function NoteGrid (props: NoteGridProps) {
                 noteTitleFont={props.noteTitleFont}
                 noteContentFontSize={props.noteContentFontSize}
                 noteTitleFontSize={props.noteTitleFontSize}
+                richTextEditorEnabled={props.richTextEditorEnabled}
                 noteSize={props.noteSize}
                 showNoteTitles={props.showNoteTitles}
                 showNoteFooters={props.showNoteFooters}