+++ src/components/SortableNote.tsx
@@ -24,6 +24,7 @@ type SortableNoteProps = {
   noteTitleFont: NoteFontPreference;
   noteContentFontSize: NoteFontSize;
   noteTitleFontSize: NoteFontSize;
+  richTextEditorEnabled: boolean;
   noteSize: NoteSizePreference;
   showNoteTitles: boolean;
   showNoteFooters: boolean;
@@ -68,6 +69,7 @@ function SortableNote(props: SortableNoteProps) {
         noteTitleFont={props.noteTitleFont}
         noteContentFontSize={props.noteContentFontSize}
         noteTitleFontSize={props.noteTitleFontSize}
+        richTextEditorEnabled={props.richTextEditorEnabled}
         noteSize={props.noteSize}
         showNoteTitles={props.showNoteTitles}
         showNoteFooters={props.showNoteFooters}